Add useQuery to welcomePageSearch, refactor
