+++ src/containers/WelcomePage/WelcomePage.jsx
@@ -6,18 +6,11 @@
  *
  */
 
-import React, { useState, Fragment } from 'react';
+import React, { Fragment } from 'react';
 import { HelmetWithTracker } from '@ndla/tracker';
 import PropTypes from 'prop-types';
-import {
-  FrontpageHeader,
-  FrontpageFilm,
-  OneColumn,
-  FrontpageSearch,
-} from '@ndla/ui';
+import { FrontpageHeader, FrontpageFilm, OneColumn } from '@ndla/ui';
 import { injectT } from '@ndla/i18n';
-import { Query } from 'react-apollo';
-import debounce from 'lodash.debounce';
 import Spinner from '@ndla/ui/lib/Spinner';
 import {
   GraphQLFrontpageShape,
@@ -31,50 +24,13 @@ import FrontpageSubjects from './FrontpageSubjects';
 import { FILM_PAGE_PATH, ALLOWED_SUBJECTS } from '../../constants';
 import SocialMediaMetadata from '../../components/SocialMediaMetadata';
 import config from '../../config';
-import handleError from '../../util/handleError';
-import { frontpageSearchQuery } from '../../queries';
 
-import { topicsNotInNDLA } from '../../util/topicsHelper';
-import { mapSearchToFrontPageStructure } from '../../util/searchHelpers';
-import { toSearch } from '../../routeHelpers';
 import { getLocaleUrls } from '../../util/localeHelpers';
 import { LocationShape } from '../../shapes';
 import BlogPosts from './BlogPosts';
-import { searchResultToLinkProps } from '../SearchPage/searchHelpers';
+import WelcomePageSearch from './WelcomePageSearch';
 
-const debounceCall = debounce(fn => fn(), 300);
 const WelcomePage = ({ t, data, loading, locale, history, location }) => {
-  const [query, setQuery] = useState('');
-  const [delayedSearchQuery, setDelayedSearchQuery] = useState('');
-  const [inputHasFocus, setInputHasFocus] = useState(false);
-  const onSearchFieldChange = query => {
-    setQuery(query);
-    debounceCall(() => setDelayedSearchQuery(query));
-  };
-
-  const allResultsUrl = toSearch(`?query=${query}`);
-
-  const onSearch = evt => {
-    evt.preventDefault();
-    history.push(allResultsUrl);
-  };
-
-  const renderInfoText = () => (
-    <span>
-      {topicsNotInNDLA.map((topic, index) => (
-        <Fragment key={topic}>
-          {index === topicsNotInNDLA.length - 1 &&
-            `${t('welcomePage.topicsConjunction')} `}
-          <strong key={topic}>
-            {topic}
-            {index < topicsNotInNDLA.length - 2 && ','}{' '}
-          </strong>
-        </Fragment>
-      ))}
-      {t('welcomePage.topicsNotAvailableFromSearch')}
-    </span>
-  );
-
   if (loading) {
     return <Spinner />;
   }
@@ -93,11 +49,6 @@ const WelcomePage = ({ t, data, loading, locale, history, location }) => {
     },
   ];
 
-  const headerMessages = {
-    searchFieldTitle: t('welcomePage.heading.messages.searchFieldTitle'),
-    menuButton: t('welcomePage.heading.messages.menuButton'),
-  };
-
   const frontPageSubjects = subjects.length > 0 && (
     <FrontpageSubjects
       subjects={subjects}
@@ -106,10 +57,6 @@ const WelcomePage = ({ t, data, loading, locale, history, location }) => {
     />
   );
 
-  const infoText =
-    topicsNotInNDLA.length > 0 && delayedSearchQuery.length > 2
-      ? renderInfoText()
-      : '';
   return (
     <Fragment>
       <HelmetWithTracker title={t('htmlTitles.welcomePage')} />
@@ -120,53 +67,11 @@ const WelcomePage = ({ t, data, loading, locale, history, location }) => {
         image={{ src: `${config.ndlaFrontendDomain}/static/logo.png` }}>
         <meta name="keywords" content={t('meta.keywords')} />
       </SocialMediaMetadata>
-
       <FrontpageHeader
         links={headerLinks}
         locale={locale}
         languageOptions={getLocaleUrls(locale, location)}>
-        <Query
-          fetchPolicy="no-cache"
-          variables={{
-            query: delayedSearchQuery,
-          }}
-          ssr={false}
-          skip={delayedSearchQuery.length <= 2}
-          query={frontpageSearchQuery}>
-          {({ data, loading, error }) => {
-            if (error) {
-              handleError(error);
-              return `Error: ${error.message}`;
-            }
-            return (
-              <FrontpageSearch
-                inputHasFocus={inputHasFocus}
-                onInputBlur={() => setInputHasFocus(false)}
-                messages={headerMessages}
-                searchFieldValue={query}
-                onSearch={onSearch}
-                onSearchFieldChange={onSearchFieldChange}
-                searchFieldPlaceholder={t(
-                  'welcomePage.heading.searchFieldPlaceholder',
-                )}
-                searchResult={
-                  delayedSearchQuery.length > 2 &&
-                  mapSearchToFrontPageStructure(
-                    data,
-                    t,
-                    delayedSearchQuery,
-                    locale,
-                  )
-                }
-                infoText={infoText}
-                onSearchInputFocus={() => setInputHasFocus(true)}
-                allResultUrl={allResultsUrl}
-                loading={loading}
-                resourceToLinkProps={searchResultToLinkProps}
-              />
-            );
-          }}
-        </Query>
+        <WelcomePageSearch history={history} locale={locale} />
       </FrontpageHeader>
       <main>
         <div data-testid="category-list">{frontPageSubjects}</div>
